MITADS-Speech/requirements.txt
BeautifulSoup4
requests
sox
progressbar2==3.47.0
pycopy-shutil